pyproject.toml
[project]
name = "nasa-space-apps"
version = "0.1.0"
description = "RAG system for space biology research papers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.23.26",
    "qdrant-client>=1.7.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "sentence-transformers>=2.3.1",
    "tiktoken>=0.5.2",
    "python-dotenv>=1.0.0",
    "langchain-huggingface>=0.3.1",
    "rich>=14.1.0",
    "langgraph>=0.6.8",
    "langchain-openai>=0.3.34",
]
